internal/plots: split plots.go into 2 (add plots_list.go)
